% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_dynamic.R
\docType{data}
\name{data_dynamic}
\alias{data_dynamic}
\title{Simulated data for the dynamic LBA model}
\format{
An object of class \code{"data.frame"}
\describe{
 \item{stim}{Code representing a stimulus shown to a subject}
 \item{repetition}{Amount of stimulus repetitions within an experimental
 block}
 \item{condition}{Whether this block contains novel or repeating stimuli}
 \item{target}{The correct response for this stimulus}
 \item{response}{The response given by the subject}
 \item{accuracy}{Whether the subject gave the correct response or not}
 \item{rt}{Reaction times (in seconds)}
 \item{block}{The block number (from 1 to 16)}
}
}
\usage{
data(data_dynamic)
}
\description{
This dataset is simulated, and contains data for the dynamic LBA model.
The main difference with data_neural.RData is that here condition, target
and block information is included since this is needed to adjust the weight
matrix. More precisely, two weight matrices are defined (one for each
condition), and hence the condition information is needed to adjust the
correct matrix.
}
\examples{

data(data_dynamic)
head(data_dynamic)

}
\references{
This data set was artificially created for the labdance package.
}
\keyword{datasets}
