% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_neural.R
\docType{data}
\name{data_neural}
\alias{data_neural}
\title{Simulated data for the standard LBA model}
\format{
An object of class \code{"data.frame"}
\describe{
 \item{stim}{Code representing a stimulus shown to a subject}
 \item{repetition}{Amount of stimulus repetitions within an experimental block}
 \item{rt}{Reaction times (in seconds)}
 \item{response}{The response given by the subject}
 \item{accuracy}{Whether the subject gave the correct response or not}
}
}
\usage{
data(data_neural)
}
\description{
This dataset is simulated, and contains data for the standard LBA model.
The main difference with data_dynamic.RData is that here condition, target
and block information are not needed to simulate reaction times and choice data.
}
\examples{

data(data_neural)
head(data_neural)

}
\references{
This data set was artificially created for the labdance package.
}
\keyword{datasets}
