% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.neural.R
\name{likelihood.neural}
\alias{likelihood.neural}
\title{likelihood.neural}
\usage{
likelihood.neural(to_optim, neural_data, conditions = NULL, netinput = NULL)
}
\arguments{
\item{to_optim}{the parameter set}

\item{neural_data}{bla}

\item{conditions}{bla}

\item{netinput}{bla}
}
\value{
A data.frame containing behavioral and/or neural data.
}
\description{
Calculates the likelihood based on the passed neural data
}
\examples{
true = param_draw(base_par = c("a", "b", "t0", "sd"), n_drift  = 8, dynamic  = F)
nLBA  = simulate.data(sub_id   = 1,
                      n_blocks  = 16,
                      true_pars = true,
                      sigma_gen = 0.01)

likelihood.neural(true, nLBA$neural, nLBA$repetition)

}
