% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.summed.R
\name{likelihood.summed}
\alias{likelihood.summed}
\title{likelihood.summed}
\usage{
likelihood.summed(
  to_optim,
  rt,
  response,
  neural_data = NULL,
  conditions = NULL,
  wr = NULL,
  netinput = NULL,
  sigma_mod = NULL
)
}
\arguments{
\item{to_optim}{bla}

\item{rt}{bla}

\item{response}{bla}

\item{neural_data}{bla}

\item{conditions}{bla}

\item{wr}{bla}

\item{netinput}{bla}

\item{sigma_mod}{bla}
}
\value{
A data.frame containing behavioral and/or neural data.
}
\description{
Returns sum of likelihoods based on neural data and behavioral data
}
\examples{
true = param_draw(base_par = c("a", "b", "t0", "sd"), n_drift  = 8, dynamic  = F)
nLBA  = simulate.data(sub_id   = 1,
                      n_blocks  = 16,
                      true_pars = true,
                      sigma_gen = 0.01)

likelihood.summed(to_optim    = true,
                  rt          = nLBA$rt,
                  response    = nLBA$response,
                  neural_data = nLBA$neural,
                  conditions  = nLBA$repetition,
                  sigma_mod   = 0.01)

}
