% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_behavioral.R
\name{likelihood_behavioral}
\alias{likelihood_behavioral}
\title{likelihood_behavioral}
\usage{
likelihood_behavioral(to_optim, dataset = NULL)
}
\arguments{
\item{to_optim}{The likelihood for parameter set to_optim is calculated.}

\item{dataset}{Optional: only when empirical data is available.
This allows data to be generated relying on stimuli actually seen by
subjects.}
}
\value{
numeric value indicating the likelihood of a parameter set given
    the available behavioral data.
}
\description{
Calculates the -log(likelihood) based on the passed behavioral
    (i.e. reaction times and choice) data
}
\examples{
require(labdance)

set.seed(2022)

# get parameters, simulate data and calculate likelihood
true = param_draw(base_par = c("a", "b", "t0", "sd"),
                  n_drift  = 8,
                  dynamic  = F)
simulated = simulate_data(true_pars = true,
                          dataset   = NULL)
ll.true = likelihood_behavioral(to_optim = true,
                                dataset  = simulated)

# calculate likelihood for another parameter set
test = param_draw(base_par = c("a", "b", "t0", "sd"),
                  n_drift  = 8,
                  dynamic  = F)
ll.test = likelihood_behavioral(to_optim = test,
                                dataset  = simulated)

# check that likelihood is lowest for the true parameter set
sprintf("Negative loglikelihood for true parameters: \%.02f", ll.true)
# [1] "Negative loglikelihood for true parameters: 747.29"
sprintf("Negative loglikelihood for other parameters: \%.02f", ll.test)
# [1] "Negative loglikelihood for other parameters: 926.55"


}
