% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_neural.R
\name{likelihood_neural}
\alias{likelihood_neural}
\title{likelihood_neural}
\usage{
likelihood_neural(to_optim, dataset = NULL)
}
\arguments{
\item{to_optim}{The likelihood for parameter set to_optim is calculated.}

\item{dataset}{Optional: only when empirical data is available.
This allows data to be generated relying on stimuli actually seen by
subjects.}
}
\value{
numeric value indicating the likelihood of a parameter set given
    the neural (or model output) data.
}
\description{
Helper function: called when "likelihood.summed" is called.
    Calculates the likelihood given passed (neural) data.
}
\examples{
require(labdance)

set.seed(2022)

# neural LBA
true = param_draw(base_par = c("a", "b", "t0", "sd"),
                  n_drift  = 8,
                  dynamic  = FALSE)
dataset = simulate_data(true_pars = true,
                        dataset   = NULL,
                        sigma_gen = 0.01)
likelihood_neural(to_optim = true,
                  dataset  = dataset)
# [1] 0.04932333

# dynamic neural LBA
true = param_draw(base_par = c("a", "b", "t0", "sd", "beta"),
                  n_drift  = NULL,
                  dynamic  = TRUE)
dataset = simulate_data(true_pars = true,
                        dataset   = NULL,
                        sigma_gen = 0.01)
likelihood_neural(to_optim = true,
                  dataset  = dataset)
# [1] 0.05276045

}
