% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negloglik.behavioral.R
\name{negloglik.behavioral}
\alias{negloglik.behavioral}
\title{negloglik.behavioral}
\usage{
negloglik.behavioral(to_optim, rt, response, conditions = NULL, wr = NULL)
}
\arguments{
\item{to_optim}{the parameter set}

\item{rt}{bla}

\item{response}{bla}

\item{conditions}{bla}

\item{wr}{bla}
}
\value{
A data.frame containing behavioral and/or neural data.
}
\description{
Calculates the -log(likelihood) based on the passed behavioral (i.e. reaction times and
    choice) data
}
\examples{
true = param_draw(base_par = c("a", "b", "t0", "sd"),
                  n_drift  = 8,
                  dynamic  = F)
LBA  = simulate.data(1, 16, true)

negloglik.behavioral(to_optim   = true,
                     rt         = LBA$rt,
                     response   = LBA$response,
                     conditions = LBA$repetition)

}
