% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netinputs.R
\name{netinputs}
\alias{netinputs}
\title{netinputs}
\usage{
netinputs(beta, wr)
}
\arguments{
\item{beta}{bla}

\item{wr}{bla}
}
\value{
A data.frame.
}
\description{
Generates how the netinputs at the output level evolve over time as a function of the learning rate
    beta and the weight resets
}
\examples{
ni = netinputs(beta = 0.5, wr = c(rep(0, times = 31), 1))

}
