% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.data.R
\name{simulate.data}
\alias{simulate.data}
\title{simulate.data}
\usage{
\method{simulate}{data}(sub_id = NULL, n_blocks = NULL, true_pars, sigma_gen = NULL)
}
\arguments{
\item{sub_id}{bla}

\item{n_blocks}{bla}

\item{true_pars}{bla}

\item{sigma_gen}{bla}
}
\value{
A data.frame containing behavioral and/or neural data.
}
\description{
Function that makes calls to simulate.dynamic() or simulate.neural() depending on the type of
    data that is needed.
}
\examples{
true = param_draw(dynamic  = T)
dnLBA = simulate.data(true_pars = true, sigma_gen = 0.01)

}
