% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.dynamic.R
\name{simulate.dynamic}
\alias{simulate.dynamic}
\title{simulate.dynamic}
\usage{
\method{simulate}{dynamic}(true_pars, sigma_gen = NULL)
}
\arguments{
\item{true_pars}{bla}

\item{sigma_gen}{bla}
}
\value{
A data.frame containing behavioral and/or neural data.
}
\description{
Helper function called by simulate.data() when generating data.
}
\examples{
true = param_draw(dynamic = T)
dLBA = simulate.data(true_pars = true)

}
