% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.neural.R
\name{simulate.neural}
\alias{simulate.neural}
\title{simulate.neural}
\usage{
\method{simulate}{neural}(sub_id, n_blocks, true_pars, sigma_gen = NULL)
}
\arguments{
\item{sub_id}{bla}

\item{n_blocks}{bla}

\item{true_pars}{bla}

\item{sigma_gen}{bla}
}
\value{
A data.frame containing behavioral and neural data.
}
\description{
Helper function called by simulate.data() when generating data.
}
\examples{
true = param_draw(base_par = c("a", "b", "t0", "sd"),
                  n_drift  = 8,
                  dynamic  = F)
nLBA  = simulate.data(sub_id    = 1,
                      n_blocks  = 16,
                      true_pars = true,
                      sigma_gen = 0.01)
}
