% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negloglik.behavioral.R
\name{negloglik.behavioral}
\alias{negloglik.behavioral}
\title{negloglik.behavioral}
\usage{
negloglik.behavioral(to_optim, dataset)
}
\arguments{
\item{to_optim}{The likelihood for parameter set to_optim is calculated.}

\item{dataset}{Bla bla}
}
\value{
numeric value indicating the likelihood of a parameter set given
    the available behavioral data.
}
\description{
Calculates the -log(likelihood) based on the passed behavioral
    (i.e. reaction times and choice) data
}
\examples{
true = param_draw(base_par = c("a", "b", "t0", "sd"),
                  n_drift  = 8,
                  dynamic  = F)
LBA  = simulate.data(1, 16, true)

negloglik.behavioral(to_optim   = true,
                     rt         = LBA$rt,
                     response   = LBA$response,
                     conditions = LBA$repetition)
# [1] -423.1392

}
