% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netinputs.R
\name{netinputs}
\alias{netinputs}
\title{netinputs}
\usage{
netinputs(beta = 0.5, dataset = NULL)
}
\arguments{
\item{beta}{The learning rate. Influences the change in model weights. Since
the netinput is the dot product of the input activation and the weights,
beta has a direct impact of the level of activation at the output units.}

\item{dataset}{Optional: only when empirical data is available.
Replaces the stimuli and condition info by empirical data. Hence,
this allows data to be generated relying on stimuli actually seen by
subjects.}
}
\value{
list consisting of M elements (the number of output units), where each
    subset consists of N elements (the number of trials).
}
\description{
Generates how the netinputs at the output level of a dynamic
    model evolve over time as a function of the learning rate beta and
    the weight resets. Since our dynamic model has two output units, netinput
    has N times 2 values, where N denotes the number of trials.
}
\examples{
require(labdance)

# load prepared empirical data
data("data_dynamic")

# simulate netinputs based on data and low beta
ni = netinputs(beta = .2,
               dataset = data_dynamic)
print(rbind(ni[[1]][20:25], ni[[2]][20:25]))

#            [,1]      [,2]      [,3]      [,4]      [,5]      [,6]
# [1,] 0.4082713 0.4082713 0.5917287 0.6112978 0.6296023 0.3887022
# [2,] 0.5917287 0.5917287 0.4082713 0.3887022 0.3703977 0.6112978

# simulate netinputs based on data and high beta
ni = netinputs(beta = .7,
               dataset = data_dynamic)
print(rbind(ni[[1]][20:25], ni[[2]][20:25]))

#            [,1]      [,2]      [,3]      [,4]     [,5]      [,6]
# [1,] 0.2523346 0.2523346 0.7476654 0.7795135 0.804895 0.2204865
# [2,] 0.7476654 0.7476654 0.2523346 0.2204865 0.195105 0.7795135

}
