% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recovery.R
\name{recovery}
\alias{recovery}
\title{recovery}
\usage{
recovery(base_par, df, cycles, sigma_gen = NULL, sigma_mod = NULL)
}
\arguments{
\item{base_par}{bla}

\item{df}{bla}

\item{cycles}{bla}

\item{sigma_gen}{bla}

\item{sigma_mod}{bla}
}
\value{
A data.frame containing behavioral and/or neural data.
}
\description{
Returns sum of likelihoods based on neural data and behavioral data
}
\examples{
# simulate data based on known parameters
true = param_draw(base_par = c("a", "b", "t0", "sd"), n_drift  = 8, dynamic  = F)
nLBA  = simulate.data(sub_id    = 1,
                      n_blocks  = 16,
                      true_pars = true,
                      sigma_gen = 0.01)

recovery = recovery(base_par  = c("a", "b", "t0", "sd"),
                    df        = nLBA,
                    cycles    = 500,
                    sigma_mod = 0.01)

}
