% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.dynamic.R
\name{simulate.dynamic}
\alias{simulate.dynamic}
\title{simulate.dynamic}
\usage{
simulate.dynamic(true_pars = true, n_blocks  = 16, dataset = d)
}
\arguments{
\item{n_blocks}{Determines the number of experimental blocks within a dataset.}

\item{true_pars}{The LBA parameters used to generate the data.}

\item{sigma_gen}{Optional: only for neural models.
The variance in the generation of the true neural data.}

\item{dataset}{Optional: only when empirical data is available.
Replaces the stimuli and conditions by information observed in the data. Hence,
this allows data to be generated relying on stimuli actually seen by
subjects.}
}
\value{
data.frame containing behavioral and/or neural data.
}
\description{
Helper function called by simulate.data() whenever data from
    a dynamic model (dLBA, dnLBA) needs to be generated.
}
\examples{
require(labdance)

set.seed(2022)

# load prepared empirical data
load("data/simulate.dynamic.RData")

# get dLBA parameters
true = param.draw(base_par = c("a", "b", "t0", "sd", "beta"),
                  n_drift  = NULL,
                  dynamic  = TRUE)
# simulate data retaining the stimulus order shown to subject 2
simulated = simulate.dynamic(true_pars = true,
                             n_blocks  = 16,
                             dataset   = d)
head(simulated)

#   stim target condition       rt response   mean_v1   mean_v2
# 1    1      1     novel 1.807239        2 0.5000000 0.5000000
# 2    1      1     novel 4.307053        2 0.5280923 0.4719077
# 3    3      2     novel 1.107353        1 0.5000000 0.5000000
# 4    4      2     novel 3.858292        1 0.5000000 0.5000000
# 5    2      1     novel 1.873553        2 0.5000000 0.5000000
# 6    3      2     novel 2.170866        2 0.4719077 0.5280923

}
