% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.neural.R
\name{simulate.neural}
\alias{simulate.neural}
\title{simulate.neural}
\usage{
\method{simulate}{neural}(sub_id = 1, n_blocks = 16, true_pars = NULL, sigma_gen = 0.01)
}
\arguments{
\item{sub_id}{The subject identifier. Can be used to generate different datasets
that each belong to a different "subject".}

\item{n_blocks}{Determines the number of experimental blocks within a dataset.}

\item{true_pars}{The LBA parameters used to generate the data.}

\item{sigma_gen}{Optional: only for neural models.
The variance in the generation of the true neural data.}
}
\value{
data.frame containing behavioral and neural data.
}
\description{
Helper function called by simulate.data() whenever data from
    a neural model (nLBA, dnLBA) needs to be generated.
}
\examples{
true = param_draw(base_par = c("a", "b", "t0", "sd"),
                  n_drift  = 8,
                  dynamic  = F)
simulated  = simulate.data(sub_id    = 1,
                           n_blocks  = 16,
                           true_pars = true,
                           sigma_gen = 0.01)
head(simulated)

#     stim repetition       rt response block_nr    neural subject
# 1      1          1 1.801657        2        1 0.4283653       1
# 2      2          1 2.364796        2        1 0.4475578       1
# 3      3          1 1.465132        2        1 0.4494453       1
# 4      4          1 1.373814        2        1 0.4271550       1
# 5      1          2 1.387029        2        1 0.4611560       1
# 6      2          2 1.150407        2        1 0.4488713       1

}
